Add back the basic mandelbrot example
