+++ examples/mandelbrot-interactive/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "mandelbrot"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+compute = { path = "../.." }
+
+anyhow = "1.0.95"
+encase = { version = "0.10.0", features = ["nalgebra"] }
+image = "0.25.5"
