+++ examples/mandelbrot-interactive/src/compute.wgsl
