+++ examples/mandelbrot-interactive/src/render.wgsl
