+++ examples/mandelbrot/src/main.rs
@@ -1,92 +1,54 @@
 use anyhow::Result;
 
 use compute::{
-    buffer::UniformBuffer,
     export::{
-        egui::{self, Context, Slider},
         encase::ShaderType,
         nalgebra::{Vector2, Vector3},
-        wgpu::{include_wgsl, RenderPass, ShaderStages},
-        winit::{dpi::LogicalSize, window::WindowAttributes},
+        wgpu::include_wgsl,
     },
     gpu::Gpu,
-    interactive::Interactive,
-    pipeline::{compute::ComputePipeline, render::RenderPipeline},
 };
+use image::{ImageBuffer, Rgb};
 
-const SIZE: Vector2<u32> = Vector2::new(4096, 4096);
-
-#[derive(ShaderType, Default)]
+#[derive(ShaderType)]
 struct Uniform {
     size: Vector2<u32>,
     zoom: f32,
 }
 
-struct App {
-    compute: ComputePipeline,
-    uniform: UniformBuffer<Uniform>,
-    render: RenderPipeline,
-
-    zoom: f32,
-}
-
-impl Interactive for App {
-    fn render(&mut self, _gpu: &Gpu, render_pass: &mut RenderPass) {
-        self.uniform
-            .upload(Uniform {
-                size: SIZE,
-                zoom: self.zoom,
-            })
-            .unwrap();
-        self.compute
-            .dispatch(Vector3::new(SIZE.x / 8, SIZE.y / 8, 1));
-        self.render.draw_screen_quad(render_pass);
-    }
-
-    fn ui(&mut self, ctx: &Context) {
-        ctx.input(|input| self.zoom += input.smooth_scroll_delta.y / 100.0);
-
-        egui::Window::new("Mandelbrot").show(ctx, |ui| {
-            ui.horizontal(|ui| {
-                ui.label("Zoom");
-                ui.add(Slider::new(&mut self.zoom, 0.0..=12.0))
-            });
-        });
-    }
-}
+const SIZE: Vector2<u32> = Vector2::new(4096, 4096);
 
 fn main() -> Result<()> {
     let gpu = Gpu::init()?;
 
-    let uniform = gpu.create_uniform(Uniform::default()).unwrap();
-    let buffer = gpu
-        .create_storage(vec![0; (SIZE.x * SIZE.y) as usize])
-        .unwrap();
+    let uniform = gpu.create_uniform(Uniform {
+        size: SIZE,
+        zoom: 0.0,
+    })?;
+    let buffer = gpu.create_storage(vec![0; (SIZE.x * SIZE.y) as usize])?;
 
-    let compute = gpu
-        .compute_pipeline(include_wgsl!("compute.wgsl"))
+    let pipeline = gpu
+        .compute_pipeline(include_wgsl!("shader.wgsl"))
         .bind_buffer(&uniform)
         .bind_buffer(&buffer)
         .finish();
 
-    let render = gpu
-        .render_pipeline(include_wgsl!("render.wgsl"))
-        .bind_buffer(&uniform, ShaderStages::FRAGMENT)
-        .bind_buffer(&buffer, ShaderStages::FRAGMENT)
-        .finish();
+    for zoom in 0..10_0 {
+        uniform.upload(Uniform {
+            size: SIZE,
+            zoom: zoom as f32 / 10.0,
+        })?;
 
-    gpu.create_window(
-        WindowAttributes::default()
-            .with_title("Mandelbrot")
-            .with_inner_size(LogicalSize::new(1920, 1080)),
-        App {
-            compute,
-            uniform,
-            render,
-            zoom: 0.0,
-        },
-    )
-    .run()?;
+        pipeline.dispatch(Vector3::new(SIZE.x / 8, SIZE.y / 8, 1));
+        buffer.download_async(move |result| {
+            ImageBuffer::from_par_fn(SIZE.x, SIZE.y, |x, y| {
+                let color = result[(y * SIZE.x + x) as usize];
+                Rgb([color as u8, (color >> 8) as u8, (color >> 16) as u8])
+            })
+            .save(format!("rec/out-{zoom:0>4}.png"))
+            .unwrap();
+        });
+    }
 
     Ok(())
 }
